# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Microservice to manage PagoPA Backoffice",
    "termsOfService" : "https://www.pagopa.gov.it/",
    "title" : "SelfCare Backoffice",
    "version" : "1.4.5"
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  }, {
    "url" : "https://{host}{basePath}",
    "variables" : {
      "basePath" : {
        "default" : "/backoffice/v1"
      },
      "host" : {
        "default" : "api.dev.platform.pagopa.it",
        "enum" : [ "api.dev.platform.pagopa.it", "api.uat.platform.pagopa.it", "api.platform.pagopa.it" ]
      }
    }
  } ],
  "paths" : {
...
    "/configurations/payment-types" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getPaymentTypes",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentTypes"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get list of payment type",
        "tags" : [ "Configuration" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/configurations/wfesp-plugins" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getWfespPlugins",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WfespPluginConfs"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Update a WrapperChannel on Cosmodb",
        "tags" : [ "Configuration" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/configurations/wrapper/{wrapper-type}/status/{wrapper-status}" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getWrapperByTypeAndStatus",
        "parameters" : [ {
          "description" : "Type of Wrapper like CHANNEL or STATION",
          "in" : "path",
          "name" : "wrapper-type",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CHANNEL", "STATION" ]
          }
        }, {
          "description" : "Validation Status of a CHANNEL or STATION",
          "in" : "path",
          "name" : "wrapper-status",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "APPROVED", "TO_CHECK", "TO_FIX", "TO_CHECK_UPDATE", "TO_FIX_UPDATE" ]
          }
        }, {
          "description" : "Number of elements on one page. Default = 50",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "description" : "Page number. Page value starts from 0",
          "in" : "query",
          "name" : "page",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Broker code filter for search",
          "in" : "query",
          "name" : "broker-code",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Query with sql like parameter for field id search",
          "in" : "query",
          "name" : "id-like",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Method of sorting",
          "in" : "query",
          "name" : "sorting",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WrapperEntitiesList"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get Wrapper Channel Details from cosmos db",
        "tags" : [ "Configuration" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "PageInfo" : {
        "type" : "object",
        "properties" : {
          "items_found" : {
            "type" : "integer",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total_items" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total_pages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "PaymentType" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "payment_type" : {
            "type" : "string"
          }
        }
      },
      "PaymentTypes" : {
        "type" : "object",
        "properties" : {
          "payment_types" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PaymentType"
            }
          }
        }
      },
...
      "WfespPluginConf" : {
        "required" : [ "id_serv_plugin" ],
        "type" : "object",
        "properties" : {
          "id_bean" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "id_serv_plugin" : {
            "maxLength" : 35,
            "minLength" : 0,
            "type" : "string"
          },
          "pag_const_string_profile" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "pag_rpt_xpath_profile" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "pag_soap_rule_profile" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "WfespPluginConfs" : {
        "required" : [ "wfesp_plugin_confs" ],
        "type" : "object",
        "properties" : {
          "wfesp_plugin_confs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WfespPluginConf"
            }
          }
        }
      },
...
      "WrapperEntitiesList" : {
        "required" : [ "page_info", "wrapper_entities" ],
        "type" : "object",
        "properties" : {
          "page_info" : {
            "$ref" : "#/components/schemas/PageInfo"
          },
          "wrapper_entities" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WrapperEntitiesObject"
            }
          }
        }
      },
      "WrapperEntitiesObject" : {
        "type" : "object",
        "properties" : {
          "brokerCode" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdBy" : {
            "type" : "string"
          },
          "entities" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WrapperEntityObject"
            }
          },
          "id" : {
            "type" : "string"
          },
          "modifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "modifiedBy" : {
            "type" : "string"
          },
          "modifiedByOpt" : {
            "type" : "string"
          },
          "new" : {
            "type" : "boolean"
          },
          "note" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "APPROVED", "TO_CHECK", "TO_FIX", "TO_CHECK_UPDATE", "TO_FIX_UPDATE" ]
          },
          "type" : {
            "type" : "string",
            "enum" : [ "CHANNEL", "STATION" ]
          }
        }
      },
...
      "WrapperEntityObject" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "entity" : {
            "type" : "object"
          },
          "id" : {
            "type" : "string"
          },
          "modifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "modifiedBy" : {
            "type" : "string"
          },
          "modifiedByOpt" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "APPROVED", "TO_CHECK", "TO_FIX", "TO_CHECK_UPDATE", "TO_FIX_UPDATE" ]
          },
          "type" : {
            "type" : "string",
            "enum" : [ "CHANNEL", "STATION" ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "JWT" : {
        "bearerFormat" : "JWT",
        "description" : "JWT token get after Login",
        "scheme" : "bearer",
        "type" : "http"
      },
      "SubKey" : {
        "description" : "The Azure Subscription Key to access this API.",
        "in" : "header",
        "name" : "Ocp-Apim-Subscription-Key",
        "type" : "apiKey"
      }
    }
  }
}